recipes-bsp: Use firmware files from linux-firmware packages
TI is now pushing updated firmware files to upstream linux-firmware
and Broadcom / Cypress is no longer updating the BCM* firmware files.
We now just need to provide the configuration files necessary to
provide an optimal wireless experience on our hardware.

This also fixes file conflicts that are happening due to the inclusion
of files in linux-firmware.

Signed-off-by: Jon Nettleton <jon@solid-run.com>